My Pyrometric file
The Pymetrics Test is an online assessment in which candidates have 25 minutes to play through a series of twelve mini-games that take 2 to 3 minutes each. The assessment uses data science and algorithms to assess a broad range of 91 different cognitive traits. It accesses candidates on the basis of attention, decision making, effort, emotion, fairness, focus, generosity,risk-tolerance and learning based on years of psychological test and AI methods.